requirements.txt
mouse
keyboard
pyserial